feat(#1397): updated root layout
